feat(DataBar)!: remove top-level sub-component(s)
