--- src/components/DataBarSegment/index.ts
@@ -1,2 +1,0 @@
-export { DataBarSegment as default } from './DataBarSegment';
-export type { Props, Variants } from './DataBarSegment';
